+++ libraries/exoplayer/src/main/java/androidx/media3/exoplayer/ExoPlayerImpl.java
@@ -98,9 +98,11 @@ import androidx.media3.exoplayer.analytics.MediaMetricsListener;
 import androidx.media3.exoplayer.analytics.PlayerId;
 import androidx.media3.exoplayer.audio.AudioRendererEventListener;
 import androidx.media3.exoplayer.metadata.MetadataOutput;
+import androidx.media3.exoplayer.source.MaskingMediaSource;
 import androidx.media3.exoplayer.source.MediaSource;
 import androidx.media3.exoplayer.source.MediaSource.MediaPeriodId;
 import androidx.media3.exoplayer.source.ShuffleOrder;
+import androidx.media3.exoplayer.source.TimelineWithUpdatedMediaItem;
 import androidx.media3.exoplayer.source.TrackGroupArray;
 import androidx.media3.exoplayer.text.TextOutput;
 import androidx.media3.exoplayer.trackselection.ExoTrackSelection;
@@ -753,6 +755,11 @@ import java.util.concurrent.TimeoutException;
       return;
     }
     toIndex = min(toIndex, playlistSize);
+    if (canUpdateMediaSourcesWithMediaItems(fromIndex, toIndex, mediaItems)) {
+      // Update MediaSources directly without creating new ones if possible.
+      updateMediaSourcesWithMediaItems(fromIndex, toIndex, mediaItems);
+      return;
+    }
     List<MediaSource> mediaSources = createMediaSources(mediaItems);
     if (mediaSourceHolderSnapshots.isEmpty()) {
       // Handle initial items in a playlist as a set operation to ensure state changes and initial
@@ -1938,7 +1945,7 @@ import java.util.concurrent.TimeoutException;
         List<Timeline> timelines = ((PlaylistTimeline) newTimeline).getChildTimelines();
         checkState(timelines.size() == mediaSourceHolderSnapshots.size());
         for (int i = 0; i < timelines.size(); i++) {
-          mediaSourceHolderSnapshots.get(i).timeline = timelines.get(i);
+          mediaSourceHolderSnapshots.get(i).updateTimeline(timelines.get(i));
         }
       }
       boolean positionDiscontinuity = false;
@@ -2000,7 +2007,6 @@ import java.util.concurrent.TimeoutException;
             repeatCurrentMediaItem);
     boolean mediaItemTransitioned = mediaItemTransitionInfo.first;
     int mediaItemTransitionReason = mediaItemTransitionInfo.second;
-    MediaMetadata newMediaMetadata = mediaMetadata;
     @Nullable MediaItem mediaItem = null;
     if (mediaItemTransitioned) {
       if (!newPlaybackInfo.timeline.isEmpty()) {
@@ -2011,15 +2017,14 @@ import java.util.concurrent.TimeoutException;
       }
       staticAndDynamicMediaMetadata = MediaMetadata.EMPTY;
     }
-    if (mediaItemTransitioned
-        || !previousPlaybackInfo.staticMetadata.equals(newPlaybackInfo.staticMetadata)) {
+    if (!previousPlaybackInfo.staticMetadata.equals(newPlaybackInfo.staticMetadata)) {
       staticAndDynamicMediaMetadata =
           staticAndDynamicMediaMetadata
               .buildUpon()
               .populateFromMetadata(newPlaybackInfo.staticMetadata)
               .build();
-      newMediaMetadata = buildUpdatedMediaMetadata();
     }
+    MediaMetadata newMediaMetadata = buildUpdatedMediaMetadata();
     boolean metadataChanged = !newMediaMetadata.equals(mediaMetadata);
     mediaMetadata = newMediaMetadata;
     boolean playWhenReadyChanged =
@@ -2361,7 +2366,7 @@ import java.util.concurrent.TimeoutException;
           new MediaSourceList.MediaSourceHolder(mediaSources.get(i), useLazyPreparation);
       holders.add(holder);
       mediaSourceHolderSnapshots.add(
-          i + index, new MediaSourceHolderSnapshot(holder.uid, holder.mediaSource.getTimeline()));
+          i + index, new MediaSourceHolderSnapshot(holder.uid, holder.mediaSource));
     }
     shuffleOrder =
         shuffleOrder.cloneAndInsert(
@@ -2903,6 +2908,43 @@ import java.util.concurrent.TimeoutException;
     }
   }
 
+  private boolean canUpdateMediaSourcesWithMediaItems(
+      int fromIndex, int toIndex, List<MediaItem> mediaItems) {
+    if (toIndex - fromIndex != mediaItems.size()) {
+      // Number of items doesn't match.
+      return false;
+    }
+    for (int i = fromIndex; i < toIndex; i++) {
+      MediaSource mediaSource = mediaSourceHolderSnapshots.get(i).mediaSource;
+      if (!mediaSource.canUpdateMediaItem(mediaItems.get(i - fromIndex))) {
+        return false;
+      }
+    }
+    return true;
+  }
+
+  private void updateMediaSourcesWithMediaItems(
+      int fromIndex, int toIndex, List<MediaItem> mediaItems) {
+    pendingOperationAcks++;
+    internalPlayer.updateMediaSourcesWithMediaItems(fromIndex, toIndex, mediaItems);
+    for (int i = fromIndex; i < toIndex; i++) {
+      MediaSourceHolderSnapshot snapshot = mediaSourceHolderSnapshots.get(i);
+      snapshot.updateTimeline(
+          new TimelineWithUpdatedMediaItem(snapshot.getTimeline(), mediaItems.get(i - fromIndex)));
+    }
+    Timeline newTimeline = createMaskingTimeline();
+    PlaybackInfo newPlaybackInfo = playbackInfo.copyWithTimeline(newTimeline);
+    updatePlaybackInfo(
+        newPlaybackInfo,
+        /* timelineChangeReason= */ TIMELINE_CHANGE_REASON_PLAYLIST_CHANGED,
+        /* ignored */ PLAY_WHEN_READY_CHANGE_REASON_USER_REQUEST,
+        /* ignored */ false,
+        /* ignored */ DISCONTINUITY_REASON_REMOVE,
+        /* ignored */ C.TIME_UNSET,
+        /* ignored */ C.INDEX_UNSET,
+        /* ignored */ false);
+  }
+
   private static DeviceInfo createDeviceInfo(@Nullable StreamVolumeManager streamVolumeManager) {
     return new DeviceInfo.Builder(DeviceInfo.PLAYBACK_TYPE_LOCAL)
         .setMinVolume(streamVolumeManager != null ? streamVolumeManager.getMinVolume() : 0)
@@ -2919,12 +2961,14 @@ import java.util.concurrent.TimeoutException;
   private static final class MediaSourceHolderSnapshot implements MediaSourceInfoHolder {
 
     private final Object uid;
+    private final MediaSource mediaSource;
 
     private Timeline timeline;
 
-    public MediaSourceHolderSnapshot(Object uid, Timeline timeline) {
+    public MediaSourceHolderSnapshot(Object uid, MaskingMediaSource mediaSource) {
       this.uid = uid;
-      this.timeline = timeline;
+      this.mediaSource = mediaSource;
+      this.timeline = mediaSource.getTimeline();
     }
 
     @Override
@@ -2935,6 +2979,10 @@ import java.util.concurrent.TimeoutException;
     @Override
     public Timeline getTimeline() {
       return timeline;
+    }
+
+    public void updateTimeline(Timeline timeline) {
+      this.timeline = timeline;
     }
   }