+++ libraries/test_utils/src/main/java/androidx/media3/test/utils/FakeMediaSource.java
@@ -44,6 +44,7 @@ import androidx.media3.exoplayer.source.MediaLoadData;
 import androidx.media3.exoplayer.source.MediaPeriod;
 import androidx.media3.exoplayer.source.MediaSource;
 import androidx.media3.exoplayer.source.MediaSourceEventListener;
+import androidx.media3.exoplayer.source.TimelineWithUpdatedMediaItem;
 import androidx.media3.exoplayer.source.TrackGroupArray;
 import androidx.media3.exoplayer.upstream.Allocator;
 import androidx.media3.test.utils.FakeMediaPeriod.TrackDataFactory;
@@ -98,6 +99,7 @@ public class FakeMediaSource extends BaseMediaSource {
   private final ArrayList<MediaPeriodId> createdMediaPeriods;
   private final DrmSessionManager drmSessionManager;
 
+  private boolean canUpdateMediaItems;
   private boolean preparationAllowed;
   private @MonotonicNonNull Timeline timeline;
   private boolean preparedSource;
@@ -168,6 +170,7 @@ public class FakeMediaSource extends BaseMediaSource {
     this.drmSessionManager = drmSessionManager;
     this.trackDataFactory = trackDataFactory;
     preparationAllowed = true;
+    canUpdateMediaItems = false;
   }
 
   /**
@@ -185,6 +188,15 @@ public class FakeMediaSource extends BaseMediaSource {
     }
   }
 
+  /**
+   * Sets whether the source allows to update its {@link MediaItem} via {@link #updateMediaItem}.
+   *
+   * @param canUpdateMediaItems Whether a {@link MediaItem} update is possible.
+   */
+  public void setCanUpdateMediaItems(boolean canUpdateMediaItems) {
+    this.canUpdateMediaItems = canUpdateMediaItems;
+  }
+
   @Nullable
   protected Timeline getTimeline() {
     return timeline;
@@ -196,6 +208,22 @@ public class FakeMediaSource extends BaseMediaSource {
       return FAKE_MEDIA_ITEM;
     }
     return timeline.getWindow(0, new Timeline.Window()).mediaItem;
+  }
+
+  @Override
+  public boolean canUpdateMediaItem(MediaItem mediaItem) {
+    return canUpdateMediaItems;
+  }
+
+  @Override
+  public void updateMediaItem(MediaItem mediaItem) {
+    if (timeline == null) {
+      return;
+    }
+    timeline = new TimelineWithUpdatedMediaItem(timeline, mediaItem);
+    if (preparedSource && preparationAllowed) {
+      refreshSourceInfo(timeline);
+    }
   }
 
   @Override